updating the project

+++ backend/src/services/mng.product.services.ts
@@ -1,55 +1,55 @@
-import { Document, QueryOptions } from 'mongoose';
-import { ClientSession } from 'mongoose';
+import {
+  Document,
+  FilterQuery,
+  QueryOptions,
+  ClientSession,
+  ProjectionType,
+} from 'mongoose';
 
 import { RequestError } from '../utils/globalErrorHandler';
 
-import { UsersModel } from '../models/user.model';
 import { ProductsModel } from '../models/product.model';
 import { MngProducts, MngProductsModel } from '../models/mng.product.model';
+import { BranchesModel } from '../models/branch.model';
 
 export const handleMngProductCreation = async (
   product: Partial<MngProducts> & Document,
   userId?: string,
   session?: ClientSession
 ): Promise<MngProducts> => {
-  const { productId, amount, bio } = product;
+  const { branchId, productId, quantity, bio } = product;
 
+  if (!userId) throw new RequestError('User must be registered', 400);
+  if (!branchId) throw new RequestError('Branch name must not be empty', 400);
   if (!productId) throw new RequestError('Proudct name must not be empty', 400);
-  if (!amount) throw new RequestError('Price must not be empty', 400);
+  if (!quantity) throw new RequestError('Price must not be empty', 400);
 
-  const existingUser = await UsersModel.findOne({ id: userId });
+  const newproduct = await createNewMngProduct(
+    productId,
+    branchId,
+    userId,
+    quantity,
+    bio,
+    session
+  );
 
-  const branchId = existingUser?.branchId;
-
-  if (!branchId) {
-    throw new RequestError(
-      `Can't register this branch. this branch is not existed.`,
-      500
-    );
-  } else {
-    const newproduct = await createNewMngProduct(
-      productId,
-      branchId,
-      amount,
-      bio,
-      session
-    );
-
-    return newproduct;
-  }
+  return newproduct;
 };
 
 export const createNewMngProduct = async (
   productId: string,
   branchId: string,
-  amount: number,
+  userId: string,
+  quantity: number,
   bio?: string,
   session?: ClientSession
 ): Promise<MngProducts> => {
   const newProduct = new MngProductsModel({
     productId,
     branchId,
-    amount,
+    userId,
+    quantity,
+    status: false,
     bio,
   });
 
@@ -61,11 +61,7 @@ export const handleGetMngProductsByUser = async (
   userId?: string,
   session?: ClientSession
 ): Promise<MngProducts[]> => {
-  const existingUser = await UsersModel.findOne({ id: userId });
-
-  const branchId = existingUser?.branchId;
-
-  if (!branchId) {
+  if (!userId) {
     throw new RequestError(
       `Can't register this branch. this branch is not existed.`,
       500
@@ -73,7 +69,7 @@ export const handleGetMngProductsByUser = async (
   } else {
     const products = await MngProductsModel.aggregate([
       {
-        $match: { branchId: branchId },
+        $match: { userId: userId },
       },
       {
         $lookup: {
@@ -84,14 +80,62 @@ export const handleGetMngProductsByUser = async (
         },
       },
       { $unwind: '$productDetails' },
+      {
+        $lookup: {
+          from: BranchesModel.collection.name,
+          localField: 'branchId',
+          foreignField: 'id',
+          as: 'branchDetails',
+        },
+      },
+      { $unwind: '$branchDetails' },
     ]);
     return products;
   }
+};
+
+export const handleGetMngProducts = async (
+  session?: ClientSession
+): Promise<MngProducts[]> => {
+  const products = await MngProductsModel.aggregate([
+    {
+      $lookup: {
+        from: ProductsModel.collection.name,
+        localField: 'productId',
+        foreignField: 'id',
+        as: 'productDetails',
+      },
+    },
+    { $unwind: '$productDetails' },
+    {
+      $lookup: {
+        from: BranchesModel.collection.name,
+        localField: 'branchId',
+        foreignField: 'id',
+        as: 'branchDetails',
+      },
+    },
+    { $unwind: '$branchDetails' },
+  ]);
+  return products;
 };
 
 export const handleDeleteMngProduct = async (
   id: string,
   options?: QueryOptions<MngProducts>
 ) => {
-  return await MngProductsModel.deleteOne({ id: id });
+  const product = await findOneMngProduct({ id });
+  if (product?.status) {
+    throw new RequestError('This Order could not be deleted', 400);
+  } else {
+    return await MngProductsModel.deleteOne({ id: id });
+  }
 };
+
+export async function findOneMngProduct(
+  filter?: FilterQuery<MngProducts>,
+  projection?: ProjectionType<MngProducts>,
+  options?: QueryOptions<MngProducts>
+): Promise<MngProducts | null> {
+  return await MngProductsModel.findOne(filter, projection, options);
+}
